Rework some aspects of the custom Setup.hs
This squashes a link-time bug that was appearing in downstream packages.
The main trick is to make sure we override `confHook` rather than
`postConf`, and return an updated `LocalBuildInfo` that adds the
directory containing `libblt.a` to the package description.

Hypothesis: the reason this fixes the problem is that Cabal captures and serializes
the result of the `confHook` to use in later phases.  External tools
(i.e., stack) will also read and resuse this information.  Thus, if we only
update the `PackageDescription` in-memory during our own builds, then our
updates are not propigated to downstream packages, which then fail with
linker errors.

The current head version of cabal new-build does not seem to suffer from this
problem, and I don't really understand why.